AI/requirements.txt
SpeechRecognition
google-generativeai
hackerbot
AI_utils
